requirement.txt
asgiref==3.8.1
certifi==2025.1.31
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
numpy==2.2.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
setuptools==75.8.0
sqlparse==0.5.3
wheel==0.45.1
